add configuration class in recommendation-service
