+++ microservices/recommendation-service/src/main/java/com/yavlash/microservices/core/recommendation/config/ApplicationConfiguration.java
@@ -1,0 +1,30 @@
+package com.yavlash.microservices.core.recommendation.config;
+
+import com.yavlash.microservices.core.recommendation.persistence.RecommendationEntity;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.event.ContextRefreshedEvent;
+import org.springframework.context.event.EventListener;
+import org.springframework.data.mapping.context.MappingContext;
+import org.springframework.data.mongodb.core.ReactiveMongoOperations;
+import org.springframework.data.mongodb.core.index.IndexResolver;
+import org.springframework.data.mongodb.core.index.MongoPersistentEntityIndexResolver;
+import org.springframework.data.mongodb.core.index.ReactiveIndexOperations;
+import org.springframework.data.mongodb.core.mapping.MongoPersistentEntity;
+import org.springframework.data.mongodb.core.mapping.MongoPersistentProperty;
+
+@Configuration
+@ComponentScan("com.yavlash")
+public class ApplicationConfiguration {
+    @Autowired
+    ReactiveMongoOperations mongoTemplate;
+
+    @EventListener(ContextRefreshedEvent.class)
+    public void initIndicesAfterStartup() {
+        MappingContext<? extends MongoPersistentEntity<?>, MongoPersistentProperty> mappingContext = mongoTemplate.getConverter().getMappingContext();
+        IndexResolver resolver = new MongoPersistentEntityIndexResolver(mappingContext);
+        ReactiveIndexOperations indexOps = mongoTemplate.indexOps(RecommendationEntity.class);
+        resolver.resolveIndexFor(RecommendationEntity.class).forEach(e -> indexOps.ensureIndex(e).block());
+    }
+}
